fix(B0): server-side auth redirect — no more blank screen
Root cause: the client-side useEffect redirect in the dashboard layout was
racing with React hydration, producing a persistent blank/dark screen.

Fix:
- (dashboard)/layout.tsx is now a server component — reads the accessToken
  cookie and calls redirect('/login') on the server before any HTML is sent
  to the browser. No flash, no race condition.
- ClientLayout.tsx handles logout and passes AppShell to children
- middleware.ts restored to redirect unauthenticated users to /login
  (works now because login page sets document.cookie on success)

Flow: visit / → middleware OR server layout detects no cookie → redirect
to /login (server-side, instant) → login form renders.
